requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
groq>=0.4
python-dotenv>=1.0
django-cors-headers>=4.3
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
reportlab>=4.0
gunicorn>=21.0
